Safari-609.1.11.1/LayoutTests/editing/selection/click-in-focusable-link-should-not-clear-selection-expected.txt

Ensures that the selection is not cleared when the user does a mousedown on a mouse focusable link.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS String(window.getSelection()) is "Select"
PASS String(window.getSelection()) is "Select"
PASS successfullyParsed is true

TEST COMPLETE
